Fixed numerous places where NetMessage.m_senderEndPoint wasn't set properly. ReadMessage() overrides which outputs an IPEndPoint should be much more reliable now.

+++ Lidgren.Network/NetBuffer.Read.cs
@@ -371,7 +371,7 @@ namespace Lidgren.Network
 		public IPEndPoint ReadIPEndPoint()
 		{
 			uint address = ReadUInt32();
-			int port = ReadInt32();
+			int port = (int)ReadUInt16();
 			return new IPEndPoint(new IPAddress((long)address), port);
 		}
